Limit the style cache to 1024 entries but keep the most used styles
The most used styles are never removed from the cache. Here, we use a
linkied-list for ordering the styles accordingly to the use of each one.

When the limit is reached, the less used style is removed. The less used
style is always in front (head) of the linked-list. When a style is
used, this one is always moved to the end of the linked-list.